+++ Jetelina/lib/jetelina/PgTestDBController.jl
@@ -1,5 +1,5 @@
 """
-module: TestDBController
+module: PgTestDBController
 
 Author: Ono keiji
 Version: 1.0
@@ -13,7 +13,7 @@ functions
     doSelect(sql::String)  execute ordered sql sentence on test db and acquire its speed.
     measureSqlPerformance()  execute 'select' sql sentence and write the execution speed into a file
 """
-module TestDBController
+module PgTestDBController
 
     using CSV, LibPQ, DataFrames, IterTools, Tables
     using JetelinaLog, JetelinaReadConfig
@@ -82,7 +82,8 @@ module TestDBController
             return false
         finally
             # close the connection
-            TestDBController.close_connection(conn)
+            close_connection(conn)
+#            PgTestDBController.close_connection(conn)
         end
     end
 
@@ -90,26 +91,33 @@ module TestDBController
     function measureSqlPerformance()
 
         execute 'select' sql sentence and write the execution speed into a file
+        Attention: JetelinaExperimentSqlList is created when SQLAnalyzer.main()(indeed createAnalyzedJsonFile()) runs.
+                   JetelinaExperimentSqlList does not created if there were not sql.log file and data in it.
+                   yes, this measure..() function is called after creating that file in SQLAnalyzer, but for secure. maybe too much.:p
+
     """
     function measureSqlPerformance()
         #===
-        Df_JetelinaSqlList　に格納されているsqlリストを利用するのがよさそうと思ったが、DF_Jeteli..はGenie空間にあるため、
-        web経由でないと利用できない。それだとcronとか別プロセスで利用できないので、JetelinaSqlListを開いて処理することにする。
+            Tips:
+                I know it can use Df_JetelinaSqlList here, but wanna leave a evidence what sql are executed.
+                That's reason why JetelinaExperimentSqlList file is opend here.
         ===#
         sqlFile = getFileNameFromConfigPath(JetelinaExperimentSqlList)
-        sqlPerformanceFile = getFileNameFromConfigPath(string(JetelinaSqlPerformancefile,".test"))
-        open(sqlPerformanceFile, "w") do f
-            println(f,string(JetelinaFileColumnApino,',',JetelinaFileColumnMax,',',JetelinaFileColumnMin,',',JetelinaFileColumnMean))
-            df = CSV.read( sqlFile, DataFrame )
-            for i in 1:size(df,1)
-                if startswith(df.apino[i] ,"js")
-                    p = doSelect(df.sql[i])
-                    fno::String=df.apino[i]
-                    fmax::Float64=p[1][1]
-                    fmin::Float64=p[2][1]
-                    fmean::Float64=p[3]
-                    s = """$fno,$fmax,$fmin,$fmean"""
-                    println(f,s)
+        if isfile(sqlFile)
+            sqlPerformanceFile = getFileNameFromConfigPath(string(JetelinaSqlPerformancefile,".test"))
+            open(sqlPerformanceFile, "w") do f
+                println(f,string(JetelinaFileColumnApino,',',JetelinaFileColumnMax,',',JetelinaFileColumnMin,',',JetelinaFileColumnMean))
+                df = CSV.read( sqlFile, DataFrame )
+                for i in 1:size(df,1)
+                    if startswith(df.apino[i] ,"js")
+                        p = doSelect(df.sql[i])
+                        fno::String=df.apino[i]
+                        fmax::Float64=p[1][1]
+                        fmin::Float64=p[2][1]
+                        fmean::Float64=p[3]
+                        s = """$fno,$fmax,$fmin,$fmean"""
+                        println(f,s)
+                    end
                 end
             end
         end
